+++ example/parsec/src/parsec.c
@@ -1,9 +1,8 @@
 /* -*- coding: utf-8-unix -*- */
 #include <cparsec3/parsec/setup.h>
-
-#include <cparsec3/parsec/base/runner.h>
 #include <cparsec3/parsec/types.h>
 
+#include <cparsec3/parsec/base/runner.h>
 #include <cparsec3/parsec/parser/char.h>
 #include <cparsec3/parsec/parser/choice.h>
 #include <cparsec3/parsec/parser/combinator.h>
